test: checkout: move repeated add tests
