+++ dvc/remote/base.py
@@ -699,10 +699,7 @@ class RemoteBASE(object):
             if self.protected:
                 self.protect(path_info)
             else:
-                # NOTE: we can unprotect, because `hardlink/symlink` check
-                # has been performed before, so no chance of copying,
-                # only chmod-ing.
-                self.unprotect(path_info)
+                self.unprotect(path_info, False)
 
         if progress_callback:
             progress_callback(str(path_info))
